# Improve performance of `cumsum` and `cumprod`