--- a/cupy/core/_routines_math.pxd
+++ b/cupy/core/_routines_math.pxd
@@ -15,7 +15,11 @@ cdef ndarray _ndarray_clip(ndarray self, a_min, a_max, out)
 cpdef ndarray _nansum(ndarray a, axis, dtype, out, keepdims)
 cpdef ndarray _nanprod(ndarray a, axis, dtype, out, keepdims)
 
-cdef ndarray scan(ndarray a, ndarray out=*)
+cpdef enum scan_op:
+    SCAN_SUM = 0
+    SCAN_PROD = 1
+
+cdef ndarray scan(ndarray a, op=*, dtype=*, ndarray out=*)
 cdef object _sum_auto_dtype
 cdef object _add
 cdef object _conj