… tests/cupy_tests/math_tests/test_sumprod.py