Solves #2905 

Uses the fast scan available in _math_routines and adds new kernels for batched cumsum/cumprod

Matrix size : (550, 100, 72, 50)

Previous to this PR:
```
cumsum   axis=None  :   228.304 us   +/-23.550 (min:  217.939 / max:  358.161) us  80835.688 us   +/-37.779 (min:80798.241 / max:81046.272) us
cumsum   axis=0     :   241.658 us   +/-11.215 (min:  234.609 / max:  315.020) us  84766.405 us   +/- 9.216 (min:84760.162 / max:84825.439) us
cumsum   axis=1     :   261.572 us   +/- 8.596 (min:  255.286 / max:  308.073) us  85690.466 us   +/- 6.717 (min:85684.509 / max:85727.554) us
cumsum   axis=2     :   261.069 us   +/- 6.102 (min:  255.105 / max:  292.426) us  85336.803 us   +/-49.660 (min:85323.997 / max:85682.014) us
cumsum   axis=3     :   258.408 us   +/-48.437 (min:  239.055 / max:  570.139) us  345277.179 us   +/-337.118 (min:344687.134 / max:346113.464) us
```

This PR:
```
cumsum   axis=None  :    65.940 us   +/-14.030 (min:   60.568 / max:  148.783) us  13706.375 us   +/-27.500 (min:13665.888 / max:13767.488) us
cumsum   axis=0     :    71.601 us   +/-20.592 (min:   63.724 / max:  185.169) us  53530.026 us   +/-17.431 (min:53509.281 / max:53601.856) us
cumsum   axis=1     :    64.266 us   +/- 3.725 (min:   61.088 / max:   83.882) us  37469.175 us   +/- 7.742 (min:37457.344 / max:37487.873) us
cumsum   axis=2     :    64.903 us   +/- 4.328 (min:   61.789 / max:   86.990) us  41359.191 us   +/-61.012 (min:41322.079 / max:41601.791) us
cumsum   axis=3     :    46.353 us   +/- 4.449 (min:   43.463 / max:   62.435) us  38549.706 us   +/- 9.629 (min:38543.297 / max:38612.896) us
```