@emcastillo 2020-01-16
@leofang, for a pretty large matrix I have

shape of random matrix (550, 100, 72, 50)

cumsum axis 0; original code: 293us new: 80us
cumsum axis 1; original code: 268us new: 71us
cumsum axis 2; original code: 262us new: 76us
cumsum axis 3; original code: 271 new: 51us

results are compared to numpy to verify they are correct.
I need to clean the code but it works and its blazingly fast