@leofang 2020-01-16
Thanks @emcastillo! I tested this PR and the CUB one (#2919), and this is what I got for `axis=None` (the only case that CUB can handle):

This PR:
```python
>>> # small array
>>> a = cp.random.random(1000000)
>>> print(cupyx.time.repeat(cp.cumprod, (a, )))
cumprod             :   125.646 us   +/-15.137 (min:  111.840 / max:  267.190) us    170.615 us   +/- 7.242 (min:  155.904 / max:  279.200) us
>>> print(cupyx.time.repeat(cp.cumsum, (a, )))
cumsum              :   122.630 us   +/-12.048 (min:  114.362 / max:  188.369) us    169.907 us   +/- 5.974 (min:  159.808 / max:  211.584) us
>>> 
>>> # large array
>>> a = cp.random.random((550, 100, 72, 50))
>>> print(cupyx.time.repeat(cp.cumprod, (a, )))
cumprod             :   150.219 us   +/- 7.734 (min:  139.555 / max:  397.774) us  21486.150 us   +/- 6.017 (min:21472.864 / max:21667.393) us
>>> print(cupyx.time.repeat(cp.cumsum, (a, )))
cumsum              :   321.688 us   +/-357.541 (min:  143.029 / max:16303.031) us  21578.071 us   +/-244.182 (min:21473.824 / max:37416.031) us
```
CUB:
```python
>>> # small array
>>> a = cp.random.random(1000000)
>>> print(cupyx.time.repeat(cp.cumprod, (a, )))
cumprod             :   110.483 us   +/-19.974 (min:   86.376 / max:  257.022) us    138.426 us   +/-18.373 (min:  116.416 / max:  275.616) us
>>> print(cupyx.time.repeat(cp.cumsum, (a, )))
cumsum              :   107.210 us   +/-20.455 (min:   86.410 / max:  240.864) us    136.249 us   +/-18.707 (min:  115.456 / max:  261.280) us
>>> 
>>> # large array
>>> a = cp.random.random((550, 100, 72, 50))
>>> print(cupyx.time.repeat(cp.cumprod, (a, )))
cumprod             :   167.864 us   +/-136.475 (min:   91.926 / max: 3520.521) us  11935.641 us   +/-109.447 (min:11875.840 / max:15090.272) us
>>> print(cupyx.time.repeat(cp.cumsum, (a, )))
cumsum              :   181.896 us   +/-139.883 (min:   97.750 / max: 1626.004) us  11948.364 us   +/-112.571 (min:11879.648 / max:13236.832) us
```
So it seems CUB is still faster for scanning the whole array?