@emcastillo 2020-01-17
When using plain cumsum there is some memory copy, and allocation (and type promotion) overhead to make it safe for inplace scanning, that is hindering performance compared to your code, where you are calling cub before that happens.