@emcastillo 2020-01-28
Rebased and review applied, merged the `scan` and `scan_op` functions in one `scan`.
Do you think the name is ok?
Thanks