@emcastillo 2020-02-04
Renamed some variables for a better understanding of the out returning!
Thanks for the comments